requirements.txt
flask>=2.3.0
networkx>=3.1
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
pyvis>=0.3.2
python-dotenv>=1.0.0
scikit-learn>=1.3.0
plotly>=5.18.0
chart-studio>=1.1.0
community>=1.0.0b1
python-louvain>=0.16
pyarrow>=14.0.1  # Added for parquet support
fastparquet>=2023.10.1  # Added as alternative parquet engine